+++ pipen/defaults.py
@@ -57,11 +57,18 @@ CONFIG = Diot(
     plugin_opts={},
 )
 
-CONSOLE_WIDTH: int = 100
+# Just the total width of the terminal
+# when logging with a rich.Panel()
+CONSOLE_WIDTH_WITH_PANEL: int = 100
+# The width of the terminal when the width cannot be detected,
+# we are probably logging into a file
+CONSOLE_DEFAULT_WIDTH: int = 256
 # [05/16/22 11:46:40] I
+# v0.3.4:
+# 05-16 11:11:11 I
 # The markup code is included
 # Don't modify this unless the logger formatter is changed
-CONSOLE_WIDTH_SHIFT: int = 30
+CONSOLE_WIDTH_SHIFT: int = 25
 # For pipen scheduler plugins
 SCHEDULER_ENTRY_GROUP = "pipen_sched"
 # For pipen template plugins